// generated/orval/fetch/google-health-api/models/serving.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Represents different properties and information about the serving of a specific food.
 */
export interface Serving {
  /** Optional. The number of servings. */
  amount?: number;
  /** Required. Food measurement unit */
  foodMeasurementUnit?: string;
  /** Output only. Legacy measurement unit for serving size in singular form (e.g. "piece", "gram"). */
  readonly foodMeasurementUnitDisplayName?: string;
}